requirements.txt
attrs==20.3.0
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
flask-mongoengine==1.0.0
Flask-Script==2.0.6
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mongoengine==0.21.0
py-bcrypt==0.4
pymongo==3.11.2
pyrsistent==0.17.3
python-dotenv==0.15.0
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
